Move plugins to ./plugins
