+++ .browserslistrc
@@ -13,4 +13,4 @@ iOS >= 11
 ie 11
 
 [node]
-node 22
+node 24
