+++ src/server/common/helpers/session-cache/cache-engine.js
@@ -3,7 +3,7 @@ import { Engine as CatboxMemory } from '@hapi/catbox-memory'
 
 import { createLogger } from '../logging/logger.js'
 import { buildRedisClient } from '../redis-client.js'
-import { config } from '../../../../config/config.js'
+import { config } from '#/config/config.js'
 
 export function getCacheEngine(engine) {
   const logger = createLogger()
