+++ src/server/plugins/content-security-policy.js
