+++ src/server/plugins/content-security-policy.test.js
