+++ src/server/plugins/pulse.js
@@ -1,6 +1,6 @@
 import hapiPulse from 'hapi-pulse'
 
-import { createLogger } from './logging/logger.js'
+import { createLogger } from '../common/helpers/logging/logger.js'
 
 const tenSeconds = 10 * 1000
