+++ src/server/plugins/request-tracing.js
@@ -1,6 +1,6 @@
 import { tracing } from '@defra/hapi-tracing'
 
-import { config } from '../../../config/config.js'
+import { config } from '../../config/config.js'
 
 export const requestTracing = {
   plugin: tracing.plugin,
