+++ src/server/plugins/router.js
@@ -1,9 +1,9 @@
 import inert from '@hapi/inert'
 
-import { home } from './routes/home/index.js'
-import { about } from './routes/about/index.js'
-import { health } from './routes/health/index.js'
-import { serveStaticFiles } from './common/helpers/serve-static-files.js'
+import { home } from '../routes/home/index.js'
+import { about } from '../routes/about/index.js'
+import { health } from '../routes/health/index.js'
+import { serveStaticFiles } from './serve-static-files.js'
 
 export const router = {
   plugin: {
