+++ src/server/plugins/session-cache.js
@@ -1,6 +1,6 @@
 import yar from '@hapi/yar'
 
-import { config } from '../../../../config/config.js'
+import { config } from '#/config/config.js'
 
 const sessionConfig = config.get('session')
